feat: customElements 관련 프로젝트 구조 개선
- registerCustomElements 생성
- customElements별 파일 분리
- render 메서드 추가
