+++ src/components/common/index.js
@@ -1,0 +1,2 @@
+export { PageHeader } from "./Header";
+export { PageFooter } from "./Footer";
