+++ src/components/home/index.js
@@ -1,0 +1,2 @@
+export { PostList } from "./PostList";
+export { Sidebar } from "./Sidebar";
