+++ src/registerElement.js
@@ -1,0 +1,12 @@
+import { PostList, Sidebar } from "@/components/home";
+import { PageFooter, PageHeader } from "@/components/common";
+import { BlogHome, BlogArticle } from "@/view";
+
+export default function registerCustomElements() {
+  customElements.define("page-header", PageHeader);
+  customElements.define("page-footer", PageFooter);
+  customElements.define("side-bar", Sidebar);
+  customElements.define("post-list", PostList);
+  customElements.define("blog-home", BlogHome);
+  customElements.define("blog-article", BlogArticle);
+}
