+++ src/view/index.js
@@ -1,0 +1,2 @@
+export { BlogHome } from "./Home";
+export { BlogArticle } from "./Article";
